==> 01-films-vs-books/01 Get film to book linkage.ipynb <==
# Scrape Wikipedia tables of films based on books

Parent page: https://en.wikipedia.org/wiki/Lists_of_works_of_fiction_made_into_feature_films

import wikipedia
import requests
from bs4 import BeautifulSoup
import pandas as pd
from io import StringIO

Get main page HTML

main_page_wiki = wikipedia.page("Lists of works of fiction made into feature films")
main_page_soup = BeautifulSoup(main_page_wiki.html())

Get relevant \<a\> tags

sub_page_tags = [
    tag for tag in main_page_soup.select("a") if tag.text.startswith("List of")
]

# print for sanity
for title in [tag.text for tag in sub_page_tags]:
    print(title)

Function to extract all DataFrames from a single subpage (which contains multiple tables, one per starting letter)

def collect_tables(subpage_title):
    subpage_frames = []
    subpage_title = "List of fiction works made into feature films (0–9, A–C)"
    subpage = wikipedia.page(subpage_title)
    subpage_soup = BeautifulSoup(subpage.html())
    wikitables = subpage_soup.select("table.wikitable")
    for table in wikitables:
        # convert soup object to str to represent HTML
        # pass that to pandas to extract DataFrame
        table_df = pd.read_html(StringIO(str(table)))[0]
        subpage_frames.append(table_df)

    return subpage_frames

For each page, collect tables

all_frames = []

for sub_page_tag in sub_page_tags:
    subpage_tables = collect_tables(sub_page_tag.text)
    all_frames.extend(subpage_tables)

len(all_frames)

films_to_books = pd.concat(all_frames, axis=0, ignore_index=True)
print(films_to_books.shape)
films_to_books.head()

Let's make some simplifying assumptions:

- author is always at the end of the "fiction work" after the last comma
- if we find 4 numbers + any other character (we have things like "1996-present" for book series) in brackets in either the fiction work or the film, it's the publication date
- the book/film title is everything up to the first `(` (this may not end well, who knows?)

# author
# note: in book ratings data, multiple authors are comma separated
films_to_books["author"] = (
    films_to_books["Fiction work(s)"]
    .str.split(",")
    .str[-1]
    .str.replace(" and ", ", ")
    .str.strip()
    .str.upper()
)

# publication dates
films_to_books["book_published_date"] = films_to_books["Fiction work(s)"].str.extract(
    r"\((\d{4}.*?)\)"
)
films_to_books["film_published_date"] = films_to_books[
    "Film adaptation(s)"
].str.extract(r"\((\d{4}.*?)\)")


# titles
def get_clean_title(title):
    if "(" in title:
        return title[: title.index("(")].upper().strip()
    return title.upper().strip()


films_to_books["book_title"] = films_to_books["Fiction work(s)"].apply(get_clean_title)
films_to_books["film_title"] = films_to_books["Film adaptation(s)"].apply(
    get_clean_title
)

films_to_books.sample(20, random_state=42)

films_to_books.to_parquet("./data/films_to_books.parquet", index=False)
