==> 01-films-vs-books/02 Explore book and film ratings data.ipynb <==
# Book ratings

From: https://github.com/zygmuntz/goodbooks-10k/tree/master (`books.csv`)

import pandas as pd

books = pd.read_csv("./data/raw/books.csv")
print(books.shape)
books.head()

books.head(1).T

book_ratings = books[
    [
        "goodreads_book_id",
        "authors",
        "original_publication_year",
        "original_title",
        "title",
        "average_rating",
        "small_image_url",
    ]
].copy()

book_ratings.head()

We need to normalise book ratings to /100. It's currently out of 5 so we can just multiply by 20.

book_ratings["average_rating_normalised"] = book_ratings["average_rating"] * 20

book_ratings["authors"] = book_ratings["authors"].str.upper().str.strip()
book_ratings["original_title"] = book_ratings["original_title"].str.upper().str.strip()
book_ratings["title"] = book_ratings["title"].str.upper().str.strip()

book_ratings.rename(
    columns={
        "authors": "author",
        "original_title": "book_title",
        "average_rating_normalised": "avg_book_rating",
    }
)

book_ratings.head()

book_ratings.to_parquet("./data/book_ratings.parquet", index=False)

# Films

From: https://www.kaggle.com/datasets/kashifsahil/16000-movies-1910-2024-metacritic

films = pd.read_csv("./data/raw/16k_Movies.csv")
print(films.shape)
films.head()

Normalise rating to out of 100

films["Rating"] = films["Rating"] * 10
films.head()

Extract title without extra stuff and release year

films["film_title"] = (
    films["Title"].str.replace("(re-release)", "").str.upper().str.strip()
)
films["film_published_date"] = pd.to_datetime(films["Release Date"]).dt.year
films.head(10)

(
    films.dropna(subset=["Rating"]).to_parquet(
        "./data/film_metacritic_ratings.parquet", index=False
    )
)
