==> 01-films-vs-books/03 Join datasets together.ipynb <==
import numpy as np
import pandas as pd


Base data (film to book linkage)

films_to_books = pd.read_parquet("./data/films_to_books.parquet")
films_to_books.head()

Film published date is a bit messy, take the first 4 characters as the year (some are things like "1972; Australian") and convert to float

films_to_books["film_published_date"] = (
    films_to_books["film_published_date"].str[:4].astype(float)
)
films_to_books.head()

films_to_books["book_published_date"].unique()

Delete any publication date that isn't a number

films_to_books.loc[
    films_to_books["book_published_date"].str.isnumeric() == False,
    "book_published_date",
] = np.nan

films_to_books["book_published_date"] = films_to_books["book_published_date"].astype(
    float
)

## Join book ratings data

book_ratings = pd.read_parquet("./data/book_ratings.parquet")
print(book_ratings.shape)
book_ratings.head()

Based on previous exploration, join only on title

print(len(films_to_books))

merged_book_ratings = films_to_books.merge(
    book_ratings, left_on="book_title", right_on="original_title", how="inner"
)

print(
    f"{len(merged_book_ratings)} books merged on title alone ({100 * (len(merged_book_ratings) / len(films_to_books)):.1f}%)"
)

merged_book_ratings.head()

# Films

Let's try joining on:

- title ONLY
- title AND release date

film_ratings = pd.read_parquet("./data/film_metacritic_ratings.parquet")
print(film_ratings.shape)
film_ratings.head()

print(len(films_to_books))

merged_film_ratings = films_to_books.merge(film_ratings, on="film_title", how="inner")

print(
    f"{len(merged_film_ratings)} films merged on title alone ({100 * (len(merged_film_ratings) / len(films_to_books)):.1f}%)"
)

What about adding date to disambiguate between films with the same title?

print(len(films_to_books))

merged_film_ratings_2 = films_to_books.merge(
    film_ratings, on=["film_title", "film_published_date"], how="inner"
)

print(
    f"{len(merged_film_ratings_2)} films merged on title AND date ({100 * (len(merged_film_ratings_2) / len(films_to_books)):.1f}%)"
)

Let's join Metacritic ratings only on title

print(len(merged_book_ratings))

merged_film_ratings = merged_book_ratings.merge(
    film_ratings, on="film_title", how="inner"
)

print(merged_film_ratings.shape)

merged_film_ratings.head()

final_data = (
    merged_film_ratings[
        [
            "book_title",
            "author",
            "book_published_date",
            "average_rating_normalised",
            "film_title",
            "film_published_date_y",
            "Rating",
        ]
    ]
    .drop_duplicates()
    .rename(
        columns={
            "average_rating_normalised": "avg_book_rating",
            "Rating": "avg_film_rating",
            "film_published_date_y": "film_published_date",
        }
    )
)

print(final_data.shape)
final_data.sort_values("book_title")

Quick check - any films better?

final_data[final_data["avg_film_rating"] > final_data["avg_book_rating"]]

film_ratings[film_ratings["film_title"] == "CRASH"]
