==> 01-films-vs-books/04 Explore other film review datasets.ipynb <==
`movies.csv` from https://www.kaggle.com/datasets/patkle/metacritic-scores-for-games-movies-tv-and-music

import pandas as pd

films = pd.read_csv("./data/raw/movies.csv")
print(films.shape)
films.head()

duplicate_films = films["title"].value_counts().loc[lambda x: x > 1].index
films[films.title.isin(duplicate_films)].sort_values("title")

Try this next time, it has different reviews for different versions of the same film so we could join on title AND date
